// src/types/OapDataSource.ts
export interface iOapDataSource {
  ID: string;
  createdAt?: string;
  createdBy?: string;
  modifiedAt?: string;
  modifiedBy?: string;
  dataSourceName?: string;
  dataSourceStatus?: string;
  dataFileName?: string;
  userName?: string;
  comments?: string;
  filePath?: string;
  country?: string;
  errors?: any;
}